left_offset and right_offset is integer

+++ src/Service/NavigateRange.php
@@ -61,8 +61,8 @@ class NavigateRange
     {
         if (($this->left_offset < 0 || $this->right_offset < 0) && $this->config->getTotalPages() > 1) {
             // definition of offset to the left and to the right of the selected page
-            $this->left_offset = floor(($this->config->getMaxNavigate() - 1) / 2);
-            $this->right_offset = ceil(($this->config->getMaxNavigate() - 1) / 2);
+            $this->left_offset = (int)floor(($this->config->getMaxNavigate() - 1) / 2);
+            $this->right_offset = (int)ceil(($this->config->getMaxNavigate() - 1) / 2);
 
             // adjustment, if the offset is too large left
             if ($this->config->getCurrentPage() - $this->left_offset < 1) {
